added login tokens
The basic functionality goes along the following lines:

When you login, you get a 64-char string (the login token) which is stored on
the client. Next time you login, you send that 64-char string and your username,
if the login is successful it'll send you a new string.

You need to have the same serial if you want this login to work.